Author Page Angular and Services done v1

+++ src/main/java/com/gcit/lms/dao/AuthorDAO.java
@@ -25,8 +25,23 @@ public class AuthorDAO extends BaseDAO implements ResultSetExtractor<List<Author
 		template.update("update tbl_author set authorName = ? where authorId = ?", new Object[]{author.getAuthorName(), author.getAuthorId()});
 	}
 	
-	public void deleteAuthor(Author author) throws ClassNotFoundException, SQLException{
-		template.update("delete from tbl_author where authorId = ?", new Object[] {author.getAuthorId()});
+	public void updateOrCreateAuthor(Author author) throws ClassNotFoundException, SQLException{
+		List<Author> authors = template.query("select * from tbl_author where authorId = ?", new Object[]{author.getAuthorId()}, this);
+		if(authors!=null && authors.size()>0) {
+			template.update("update tbl_author set authorName = ? where authorId = ?", new Object[]{author.getAuthorName(), author.getAuthorId()});
+		}
+		else {
+			template.update("insert into tbl_author values (?, ?)", new Object[] {author.getAuthorName(), author.getAuthorId()});
+		}
+	}
+	
+	public Author deleteAuthor(Author author) throws ClassNotFoundException, SQLException{
+		List<Author> authorRet = template.query("select * from tbl_author where authorId = ?", new Object[]{author.getAuthorId()}, this);
+		if(authorRet != null && !authorRet.isEmpty()){
+			template.update("delete from tbl_author where authorId = ?", new Object[] {author.getAuthorId()});
+			return authorRet.get(0);
+		}
+		return null;
 	}
 	
 	public List<Author> readAllAuthors(Integer pageNo) throws ClassNotFoundException, SQLException{
@@ -34,7 +49,7 @@ public class AuthorDAO extends BaseDAO implements ResultSetExtractor<List<Author
 		if(pageNo != null) {
 			setPageSize(10);
 			setPageNo(pageNo);
-			setLimit(query);
+			query = setLimit(query);
 		}
 		return template.query(query, this);
 	}
@@ -54,12 +69,14 @@ public class AuthorDAO extends BaseDAO implements ResultSetExtractor<List<Author
 	}
 	
 	public List<Author> readAuthorsByName(Integer pageNo, String  authorName) throws ClassNotFoundException, SQLException{
-		if(pageNo!=null) {
+		String query = "select * from tbl_author where authorName like ?";
+		if(pageNo != null) {
 			setPageSize(10);
 			setPageNo(pageNo);
+			query = setLimit(query);
 		}
 		authorName = "%"+authorName+"%";
-		List<Author> retList = template.query("select * from tbl_author where authorName like ?", new Object[]{authorName}, this);
+		List<Author> retList = template.query(query, new Object[]{authorName}, this);
 		return retList;
 	}
